New flow algorithm and bomb reset detection
List of changes:
- isSameDirection is now restrained if it's a pattern (67.5 vs 45).
- simulateSwingPos to extend position based on cut direction.
- findAngleViaPosition use simulateSwingPos, next note position and guide angle to find next dot direction.
- fixPatternHead modified to use new found direction.
- flowDetector find and handle individual note direction before being processed by processSwing. Also detect possible bomb reset.
- processSwing has been rewritten since it doesn't need to handle direction anymore. It still do the exact same task as before (create swing data).
- parityPrediction added bomb reset.
- techOperations to use those new methods.

Some info:
flowDetector is mostly used to preemptively find dot direction and to fix some stuff with it afterward by the use of logic.
The algorithm automatically assume that it's a bomb reset if there's a bomb in the same lane as the next note and between two notes. It's not "technically" a reset, but can be considered extra movement.

Maybe to do in the future:
Improve flow for sustained dot to follow how a player would swing at them properly, simulateSwingPos does a poor job at it, as it is very basic.